+++ Etude09_PulsesCounting/Vector3.java
@@ -1,20 +1,20 @@
 public class Vector3{
 
-	long x,y,z;
+	Float x,y,z;
 
-	public Vector3(int x, int y, int z){
+	public Vector3(Float x, Float y, Float z){
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 
 	public Vector3(){
-		this.x = 0;
-		this.y = 0;
-		this.z = 0;
+		this.x = 0.0f;
+		this.y = 0.0f;
+		this.z = 0.0f;
 	}
 	
-	public long dot(Vector3 v){
+	public Float dot(Vector3 v){
 		return (this.x * v.x) + (this.y * v.y) + (this.z * v.z);
 	}
